requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
pipp==0.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
